logout fix

+++ src/modules/wallet/components/WalletProfile.tsx
@@ -13,7 +13,11 @@ import {
   TickCircleFilled,
   Tooltip,
 } from "../../../blocks";
-import { centerMaskWalletAddress, handleCopy } from "../../../common";
+import {
+  centerMaskWalletAddress,
+  getAppParamValue,
+  handleCopy,
+} from "../../../common";
 import { useGlobalState } from "../../../context/GlobalContext";
 import { useNavigate } from "react-router-dom";
 import { useDynamicContext } from "@dynamic-labs/sdk-react-core";
@@ -37,14 +41,22 @@ const WalletProfile: FC<WalletProfileProps> = ({ selectedWallet }) => {
   const navigate = useNavigate();
 
   const handleLogOut = () => {
-    sessionStorage.removeItem("jwt");
-    dispatch({ type: "RESET_WALLET" });
-    localStorage.clear();
-    primaryWallet?.connector?.endSession();
-    dynamicLogOut();
-    handleLogOutEvent();
-    localStorage.clear();
-    navigate(APP_ROUTES.AUTH);
+    if (getAppParamValue()) {
+      // TODO: Fix this afterwards
+      window.close();
+    } else {
+      dispatch({ type: "RESET_WALLET" });
+
+      primaryWallet?.connector?.endSession();
+      dynamicLogOut();
+
+      handleLogOutEvent();
+
+      sessionStorage.removeItem("jwt");
+      localStorage.clear();
+
+      navigate(APP_ROUTES.AUTH);
+    }
   };
 
   return (
